+++ packages/blade/src/components/Box/Box.d.ts
@@ -10,6 +10,11 @@ export type BoxProps = {
   justifyContent?: CSSObject['justifyContent'];
   alignSelf?: CSSObject['alignSelf'];
   overflow?: CSSObject['overflow'];
+  position?: CSSObject['position'];
+  top?: CSSObject['top'];
+  right?: CSSObject['right'];
+  bottom?: CSSObject['bottom'];
+  left?: CSSObject['left'];
   paddingTop?: DotNotationSpacingStringToken | 'auto';
   paddingBottom?: DotNotationSpacingStringToken | 'auto';
   paddingLeft?: DotNotationSpacingStringToken | 'auto';
@@ -18,7 +23,7 @@ export type BoxProps = {
   marginBottom?: DotNotationSpacingStringToken | 'auto';
   marginLeft?: DotNotationSpacingStringToken | 'auto';
   marginRight?: DotNotationSpacingStringToken | 'auto';
-  gap?: DotNotationSpacingStringToken | 'auto';
+  gap?: DotNotationSpacingStringToken | 'auto' | number;
   minHeight?: number;
   minWidth?: number;
   maxHeight?: number;