+++ packages/blade/src/components/Button/IconButton/index.ts
@@ -1,0 +1,1 @@
+export { IconButtonProps, default as IconButton } from './IconButton';