+++ packages/blade/src/components/Checkbox/CheckboxGroup/CheckboxGroup.tsx
@@ -1,10 +1,8 @@
 import React from 'react';
-import { CheckboxGroupContent } from './CheckboxGroupContent';
 import { CheckboxGroupField } from './CheckboxGroupField';
 import { CheckboxGroupProvider } from './CheckboxGroupContext';
 import { useCheckboxGroup } from './useCheckboxGroup';
-import { FormLabelText } from '~components/Form/FormLabelText';
-import { FormHintText } from '~components/Form/FormHintText';
+import { FormLabel, FormHint } from '~components/Form';
 import Box from '~components/Box';
 
 type CheckboxGroupProps = {
@@ -104,18 +102,24 @@ const CheckboxGroup = ({
   return (
     <CheckboxGroupProvider value={contextValue}>
       <CheckboxGroupField labelledBy={ids.labelId}>
-        <FormLabelText
+        <FormLabel
+          as="span"
           neccessityIndicator={neccessityIndicator}
           position={labelPosition}
           id={ids.labelId}
           accessibillityText={accessibillityText}
         >
           {label}
-        </FormLabelText>
+        </FormLabel>
         <Box>
-          <CheckboxGroupContent>{children}</CheckboxGroupContent>
-          {showError && <FormHintText variant="error">{errorText}</FormHintText>}
-          {showHelpText && <FormHintText variant="help">{helpText}</FormHintText>}
+          <Box display="flex" flexDirection="column" gap={2}>
+            {children}
+          </Box>
+          <FormHint
+            errorText={errorText}
+            helpText={helpText}
+            type={validationState === 'error' ? 'error' : 'help'}
+          />
         </Box>
       </CheckboxGroupField>
     </CheckboxGroupProvider>