+++ packages/blade/src/components/Form/FormHintWrapper.web.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-const FormHintTextWrapper = ({ children }: { children: React.ReactNode }): React.ReactElement => {
+const FormHintWrapper = ({ children }: { children: React.ReactNode }): React.ReactElement => {
   return (
     <span style={{ display: 'flex', flexDirection: 'row', alignItems: 'center' }}>{children}</span>
   );
 };
 
-export { FormHintTextWrapper };
+export { FormHintWrapper };