+++ packages/blade/src/components/Form/FormLabel.tsx
@@ -2,11 +2,12 @@ import React from 'react';
 import { VisuallyHidden } from '~components/VisuallyHidden';
 import { Text } from '~components/Typography';
 import { BaseText } from '~components/Typography/BaseText';
-import { getPlatformType, useBreakpoint } from '~utils';
+import { getPlatformType, makeSpace, useBreakpoint } from '~utils';
 import Box from '~components/Box';
 import { useTheme } from '~components/BladeProvider';
 
-type FormLabelTextProps = {
+type CommonProps = {
+  as: 'span' | 'label';
   position?: 'top' | 'left';
   neccessityIndicator?: 'required' | 'optional' | 'none';
   accessibillityText?: string;
@@ -14,13 +15,27 @@ type FormLabelTextProps = {
   id: string;
 };
 
-const FormLabelText = ({
+type LabelProps = CommonProps & {
+  htmlFor: string;
+  as: 'label';
+};
+
+type SpanProps = CommonProps & {
+  as: 'span';
+  htmlFor?: undefined;
+};
+
+type FormLabelProps = LabelProps | SpanProps;
+
+const FormLabel = ({
+  as = 'span',
   position = 'top',
   neccessityIndicator = 'none',
   accessibillityText,
   children,
   id,
-}: FormLabelTextProps): React.ReactElement => {
+  htmlFor,
+}: FormLabelProps): React.ReactElement => {
   const { theme } = useTheme();
   const { matchedDeviceType } = useBreakpoint({ breakpoints: theme.breakpoints });
   const isDesktop = matchedDeviceType === 'desktop';
@@ -73,11 +88,11 @@ const FormLabelText = ({
       flexWrap="wrap"
     >
       <BaseText
-        lineHeight="s"
+        lineHeight={position === 'left' ? 'l' : 's'}
         fontFamily="text"
         fontWeight="bold"
         color="surface.text.subtle.lowContrast"
-        fontSize={75}
+        fontSize={position === 'left' ? 100 : 75}
       >
         {children}
         {computedAccessibilityNode}
@@ -89,16 +104,26 @@ const FormLabelText = ({
 
   // What harm can it do?
   if (isReactNative) {
-    return textNode;
+    return (
+      <Box marginRight="spacing.4" marginBottom="spacing.1">
+        {textNode}
+      </Box>
+    );
   }
 
+  const Component = as;
   // only set 120px label when device is desktop
   const width = position === 'left' && isDesktop ? '120px' : 'auto';
+
   return (
-    <span style={{ width }} id={id}>
-      {textNode}
-    </span>
+    <Component
+      htmlFor={htmlFor}
+      style={{ width, marginRight: makeSpace(theme.spacing[4]) }}
+      id={id}
+    >
+      <Box marginBottom="spacing.1">{textNode}</Box>
+    </Component>
   );
 };
 
-export { FormLabelText };
+export { FormLabel };