+++ packages/blade/src/components/Icons/ArrowRightIcon/ArrowRightIcon.web.test.tsx
@@ -1,0 +1,11 @@
+import ArrowRightIcon from '.';
+import renderWithTheme from '~src/_helpers/testing/renderWithTheme.web';
+
+describe('<ArrowRightIcon />', () => {
+  it('should render ArrowRightIcon', () => {
+    const { container } = renderWithTheme(
+      <ArrowRightIcon color="feedback.icon.neutral.lowContrast" size="large" />,
+    );
+    expect(container).toMatchSnapshot();
+  });
+});